requirements.base.txt
django >= 1.4              # Web application framework
psycopg2                   # PostgreSQL driver
south                      # Database migration service
ipython                    # Interactive Python shell
dj_database_url            # Get database URL from Heroku environment variables
pillow                     # Heroku-compatible imaging library
gunicorn                   # Production webserver
gevent                     # High speed event handling library
django-social-auth         # Authenticate using social account
django-bcrypt              # More secure password hashing
django-bootstrap-form      # Format Django forms to look nice with Twitter Bootstrap 
docutils                   # Documentation utilities
django-storages            # S3 storage for file attachments
boto                       # Python interface to Amazon Web Services
newrelic                   # Cloud-based monitoring service
django-sslify              # Require SSL site-wide
django-celery              # Distributed task queue
django-templated-email     # Template-based email
